Combine UISystem and SpriteSystem
