+++ docs/reference/interfaces/isubscriber.md
@@ -28,7 +28,6 @@
 * [TestScene](../classes/testscene.md)
 * [TestSystem](../classes/testsystem.md)
 * [TextSystem](../classes/textsystem.md)
-* [UISystem](../classes/uisystem.md)
 * [WebGLSystem](../classes/webglsystem.md)
 
 ## Index
